fix(chat): Cast temperature and max_tokens to correct types for OpenAI API
OpenAI API expects float for temperature and int for max_tokens.
Widget config from database returns strings, causing 400 Bad Request.

Fixed by explicit type casting:
- temperature: (float)
- max_tokens: (int)

Fixes error: 'Invalid type for temperature: expected a decimal, but got a string instead.'

+++ backend/app/Services/Chat/ChatOrchestrationService.php
@@ -397,8 +397,8 @@ class ChatOrchestrationService implements ChatOrchestrationServiceInterface
         $payload = [
             'model' => $request['model'] ?? ($widgetConfig['model'] ?? config('openai.chat_model', 'gpt-4o-mini')),
             'messages' => $request['messages'],
-            'temperature' => $request['temperature'] ?? ($widgetConfig['temperature'] ?? 0.2),
-            'max_tokens' => $request['max_tokens'] ?? ($widgetConfig['max_tokens'] ?? 1000),
+            'temperature' => (float) ($request['temperature'] ?? ($widgetConfig['temperature'] ?? 0.2)),
+            'max_tokens' => (int) ($request['max_tokens'] ?? ($widgetConfig['max_tokens'] ?? 1000)),
         ];
         
         // Add system prompt
